requirements.txt
Flask==0.10.1
Flask-Script==2.0.5
Flask-SQLAlchemy==2.0.0
Flask-Migrate==1.4.0
Flask-WTF==0.11.0
Flask-Login==0.2.11
Flask-Admin==1.0.9
Flask-Principal==0.4.0
Flask-Assets==0.10
Flask-Pusher==1.0.2
requests==2.7.0
WTForms-SQLAlchemy==0.1
fabric==1.10.1
fabtools==0.19.0
